+++ src/node/node.js
@@ -12,6 +12,10 @@ const {
     getPeerIdentifier, getRandomInt, getSecondsPassed, getVersion, isSameMajorVersion, getIP, createIdentifier
 } = require('./tools/utils');
 
+process.on('unhandledRejection', (reason, p) => {
+    console.log('Unhandled Rejection at: Promise', p, 'reason:', reason);
+});
+
 const DEFAULT_OPTIONS = {
     name: 'CarrIOTA Nelson',
     cycleInterval: 60,
@@ -40,7 +44,7 @@ const DEFAULT_OPTIONS = {
     lazyTimesLimit: 3, // starts to penalize peer's quality if connected so many times without new TXs
     onReady: (node) => {},
     onPeerConnected: (peer) => {},
-    onPeerRemoved: (peer) => {},
+    onPeerRemoved: (peer) => {}
 };
 
 // TODO: add node tests. Need to mock away IRI for this.
@@ -65,6 +69,15 @@ class Node extends Base {
         this.sockets = new Map();
 
         this.opts.autoStart && this.start();
+
+        // Tries to fix the issue #45 https://github.com/SemkoDev/nelson.cli/issues/45
+        // Reasoning: https://github.com/request/request/issues/2161#issuecomment-313375694
+        // Also, cleans up nelson before crashing from the sky.
+        process.on('uncaughtException', (err) => {
+            if (err.code !== 'ECONNRESET') {
+                this.end().then(() => { throw err });
+            }
+        });
     }
 
     /**
@@ -176,7 +189,7 @@ class Node extends Base {
      * @private
      */
     _getIRI () {
-        const { IRIHostname, IRIPort, IRIProtocol, silent } = this.opts;
+        const { IRIHostname, IRIPort, silent } = this.opts;
 
         return (new IRI({
             logIdent: `${this.opts.port}::IRI`,
@@ -248,6 +261,10 @@ class Node extends Base {
         this.server.on('headers', (headers) => {
             const myHeaders = this._getHeaders();
             Object.keys(myHeaders).forEach((key) => headers.push(`${key}: ${myHeaders[key]}`));
+        });
+        this.server.on('error', function (err) {
+            // basically, do nothing. Most probably a ECONNRESET error.
+            // The peer will be cleaned up on next tick.
         });
         this.log('server created...');
     }
@@ -388,13 +405,20 @@ class Node extends Base {
         ws.incoming = asServer;
         this.sockets.set(peer, ws);
 
+        ws.on('message',
+            (data) => this._addNeighbors(data, ws.incoming ? 0 : peer.data.weight)
+        );
+        ws.on('close', () => removeNeighbor('socket closed'));
+        ws.on('error', () => removeNeighbor('remotely dropped'));
+        ws.on('pong', () => { ws.isAlive = true });
+
         if (asServer) {
             onConnected().then(() => this._ready && this.iri.addNeighbors([ peer ]));
         }
         else {
-            ws.on('headers', (headers) => {
+            ws.on('upgrade', (res) => {
                 // Check for valid headers
-                const head = this._getHeaderIdentifiers(headers);
+                const head = this._getHeaderIdentifiers(res.headers);
                 if (!head) {
                     this.log('!!', 'wrong headers received', head);
                     return removeNeighbor();
@@ -412,13 +436,6 @@ class Node extends Base {
             });
             ws.on('open', onConnected);
         }
-
-        ws.on('message',
-            (data) => this._addNeighbors(data, ws.incoming ? 0 : peer.data.weight)
-        );
-        ws.on('close', () => removeNeighbor('socket closed'));
-        ws.on('error', () => removeNeighbor('remotely dropped'));
-        ws.on('pong', () => { ws.isAlive = true });
     }
 
     /**
@@ -778,9 +795,9 @@ class Node extends Base {
                 this.log(`Peer ${peer.data.hostname} (${peer.data.name}) is lazy for more than ${this.opts.lazyLimit} seconds. Removing...!`.yellow);
                 promises.push(this._removeNeighbor(peer));
             }
-            else {
+            else if (ws.readyState === 1) {
                 ws.isAlive = false;
-                ws.ping('', false, true);
+                ws.ping('', false);
             }
         });
         return Promise.all(promises).then(() => false);
@@ -850,7 +867,7 @@ class Node extends Base {
                 return this._removeNeighbors(toRemove);
             }
             return([]);
-        });
+        }).then(() => this.iri.cleanupNeighbors(Array.from(this.sockets.keys())));
     }
 
     /**